feat: separate join tx components
